requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
astroid=2.4.1=py38_0
attrs=19.3.0=py_0
beautifulsoup4=4.9.1=py38_0
blas=1.0=mkl
brotlipy=0.7.0=py38h64e0658_1000
ca-certificates=2020.4.5.2=hecda079_0
catalogue=1.0.0=py_0
certifi=2020.4.5.2=py38h32f6830_0
cffi=1.14.0=py38hc512035_1
chardet=3.0.4=py38h32f6830_1006
cryptography=2.9.2=py38hb761dab_0
cymem=2.0.3=py38hc84c608_2
cython-blis=0.4.1=py38h64e0658_1
idna=2.9=py_1
importlib-metadata=1.6.1=py38h32f6830_0
importlib_metadata=1.6.1=0
intel-openmp=2019.4=233
isort=4.3.21=py38_0
joblib=0.15.1=py_0
jsonschema=3.2.0=py38h32f6830_1
lazy-object-proxy=1.4.3=py38h1de35cc_0
libcxx=10.0.0=1
libedit=3.1.20181209=hb402a30_0
libffi=3.3=h0a44026_1
libgfortran=3.0.1=h93005f0_2
llvm-openmp=10.0.0=h28b9765_0
mccabe=0.6.1=py38_1
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.0.15=py38h5e564d8_0
mkl_random=1.1.1=py38h959d312_0
murmurhash=1.0.0=py38hc84c608_0
ncurses=6.2=h0a44026_1
numpy=1.18.1=py38h7241aed_0
numpy-base=1.18.1=py38h3304bdc_1
openssl=1.1.1g=h0b31af3_0
pandas=1.0.3=py38h6c726b0_0
pip=20.0.2=py38_3
plac=0.9.6=py_1
preshed=3.0.2=py38hc84c608_3
pycparser=2.20=py_0
pyfiglet=0.8.post1=py_0
pylint=2.5.2=py38_0
pyopenssl=19.1.0=py_1
pyrsistent=0.16.0=py38h64e0658_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.3=h26836e1_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
readline=8.0=h1de35cc_0
requests=2.23.0=pyh8c360ce_2
scikit-learn=0.22.1=py38h27c97d8_0
scipy=1.4.1=py38h44e99c9_0
setuptools=47.1.1=py38_0
six=1.15.0=py_0
soupsieve=2.0.1=py_0
spacy=2.2.4=py38ha0d09dd_1
spacy-model-en_core_web_md=2.2.5=py_0
sqlite=3.31.1=h5c1f38d_1
srsly=1.0.2=py38hc84c608_0
thinc=7.4.0=py38ha0d09dd_2
tk=8.6.8=ha441bb4_0
toml=0.10.0=pyh91ea838_0
tqdm=4.46.1=pyh9f0ad1d_0
urllib3=1.25.9=py_0
wasabi=0.6.0=py_0
wheel=0.34.2=py38_0
wrapt=1.11.2=py38h1de35cc_0
xz=5.2.5=h1de35cc_0
zipp=3.1.0=py_0
zlib=1.2.11=h1de35cc_3
